+++ projects/myrmidon/cadmus-refs-asserted-chronotope/src/public-api.ts
@@ -2,6 +2,7 @@
  * Public API Surface of cadmus-refs-asserted-chronotope
  */
 
+export type { AssertedHistoricalDate } from '@myrmidon/cadmus-refs-historical-date';
 export * from './lib/asserted-chronotope/asserted-chronotope.component';
 export * from './lib/asserted-chronotope-set/asserted-chronotope-set.component';
 export * from './lib/asserted-chronotopes.pipe';