+++ projects/myrmidon/cadmus-refs-historical-date/src/public-api.ts
@@ -8,3 +8,6 @@ export * from './lib/datation/datation.component';
 export * from './lib/historical-date/historical-date';
 export * from './lib/historical-date/historical-date.component';
 export * from './lib/historical-date.pipe';
+
+export * from './lib/asserted-historical-date/asserted-historical-date';
+export * from './lib/asserted-historical-date/asserted-historical-date.component';